[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14416 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,10 +7,10 @@ public class Main{
       BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
       StringTokenizer st = new StringTokenizer(br.readLine());
       
-      int n = Integer.parseInt(st.nextToken());
-      int m = Integer.parseInt(st.nextToken());
+      long n = Integer.parseInt(st.nextToken());
+      long m = Integer.parseInt(st.nextToken());
       
-        int result = (n+m) * (n-m);
+        long result = (n+m) * (n-m);
         bw.write(String.valueOf(result));
       bw.flush();
       bw.close();
